refactor(date-planner): share calendar between create and add-dates flows
Extract the month calendar from PlanningCreateForm into a reusable
PlanningDateCalendar with optional disabledDates. Use it in
PlanningAddDates, which is now hidden behind an "Add new dates" toggle
and disables already-proposed dates in the picker.
