+++ src/components/PlanningAddDates.tsx
@@ -5,8 +5,8 @@ import { useRouter } from "next/navigation"
 import { useLocale, useTranslations } from "next-intl"
 
 import { addPlanningDates } from "@/app/actions/planning"
+import PlanningDateCalendar from "@/components/PlanningDateCalendar"
 import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
 import { useToast } from "@/components/ui/use-toast"
 
 type PlanningAddDatesProps = {
@@ -24,7 +24,7 @@ const PlanningAddDates = ({ eventId, existingDateKeys }: PlanningAddDatesProps) 
   const router = useRouter()
   const { toast } = useToast()
 
-  const [pendingDate, setPendingDate] = useState("")
+  const [isOpen, setIsOpen] = useState(false)
   const [stagedDates, setStagedDates] = useState<string[]>([])
   const [isPending, startTransition] = useTransition()
 
@@ -39,21 +39,20 @@ const PlanningAddDates = ({ eventId, existingDateKeys }: PlanningAddDatesProps) 
     [locale]
   )
 
-  const stageDate = () => {
-    const trimmed = pendingDate.trim()
-    if (!trimmed) return
+  const toggleDate = (dateKey: string) => {
+    if (existingDateKeys.includes(dateKey)) return
 
-    if (existingDateKeys.includes(trimmed) || stagedDates.includes(trimmed)) {
-      toast({ title: t("alreadyAdded"), variant: "destructive" })
-      return
-    }
-
-    setStagedDates((previous) => sortDateKeys([...previous, trimmed]))
-    setPendingDate("")
+    setStagedDates((previous) => {
+      if (previous.includes(dateKey)) {
+        return previous.filter((existing) => existing !== dateKey)
+      }
+      return sortDateKeys([...previous, dateKey])
+    })
   }
 
-  const removeStaged = (dateKey: string) => {
-    setStagedDates((previous) => previous.filter((existing) => existing !== dateKey))
+  const cancel = () => {
+    setStagedDates([])
+    setIsOpen(false)
   }
 
   const submit = () => {
@@ -66,6 +65,7 @@ const PlanningAddDates = ({ eventId, existingDateKeys }: PlanningAddDatesProps) 
       try {
         await addPlanningDates({ eventId, dates: stagedDates })
         setStagedDates([])
+        setIsOpen(false)
         router.refresh()
       } catch (error) {
         toast({
@@ -77,42 +77,59 @@ const PlanningAddDates = ({ eventId, existingDateKeys }: PlanningAddDatesProps) 
     })
   }
 
-  return (
-    <div className="glassPanel flex flex-col gap-3">
-      <h2 className="text-xl font-semibold">{t("title")}</h2>
-      <p className="text-sm text-muted-foreground">{t("description")}</p>
-
-      <div className="flex flex-col gap-2 sm:flex-row">
-        <Input
-          type="date"
-          value={pendingDate}
-          onChange={(event) => setPendingDate(event.target.value)}
-          disabled={isPending}
-        />
-        <Button type="button" variant="outline" onClick={stageDate} disabled={isPending || !pendingDate}>
-          {t("stage")}
+  if (!isOpen) {
+    return (
+      <div className="flex">
+        <Button type="button" variant="outline" onClick={() => setIsOpen(true)}>
+          {t("openButton")}
         </Button>
       </div>
+    )
+  }
 
-      {stagedDates.length > 0 && (
-        <div className="flex flex-wrap gap-2">
-          {stagedDates.map((dateKey) => (
-            <Button
-              key={dateKey}
-              type="button"
-              variant="outline"
-              onClick={() => removeStaged(dateKey)}
-              disabled={isPending}
-            >
-              {dateFormatter.format(new Date(`${dateKey}T00:00:00.000Z`))}
-            </Button>
-          ))}
-        </div>
-      )}
+  return (
+    <div className="glassPanel flex flex-col gap-4">
+      <div className="flex flex-col gap-1">
+        <h2 className="text-xl font-semibold">{t("title")}</h2>
+        <p className="text-sm text-muted-foreground">{t("description")}</p>
+      </div>
 
-      <Button type="button" onClick={submit} disabled={isPending || stagedDates.length === 0}>
-        {isPending ? t("adding") : t("submit")}
-      </Button>
+      <PlanningDateCalendar
+        selectedDates={stagedDates}
+        onToggleDate={toggleDate}
+        disabled={isPending}
+        disabledDates={existingDateKeys}
+      />
+
+      <div className="flex flex-col gap-2">
+        <h3 className="font-semibold">{t("selectedTitle")}</h3>
+        {stagedDates.length === 0 ? (
+          <p className="text-muted-foreground">{t("empty")}</p>
+        ) : (
+          <div className="flex flex-wrap gap-2">
+            {stagedDates.map((dateKey) => (
+              <Button
+                key={dateKey}
+                type="button"
+                variant="outline"
+                onClick={() => toggleDate(dateKey)}
+                disabled={isPending}
+              >
+                {dateFormatter.format(new Date(`${dateKey}T00:00:00.000Z`))}
+              </Button>
+            ))}
+          </div>
+        )}
+      </div>
+
+      <div className="flex flex-wrap gap-2">
+        <Button type="button" onClick={submit} disabled={isPending || stagedDates.length === 0}>
+          {isPending ? t("adding") : t("submit")}
+        </Button>
+        <Button type="button" variant="outline" onClick={cancel} disabled={isPending}>
+          {t("cancel")}
+        </Button>
+      </div>
     </div>
   )
 }
