+++ src/components/PlanningCreateForm.tsx
@@ -5,17 +5,13 @@ import { useRouter } from "next/navigation"
 import { useLocale, useTranslations } from "next-intl"
 
 import { createPlanningEvent } from "@/app/actions/planning"
+import PlanningDateCalendar from "@/components/PlanningDateCalendar"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { useToast } from "@/components/ui/use-toast"
-import { cn } from "@/lib/utils"
 
 function sortDateKeys(dateKeys: string[]) {
   return [...dateKeys].sort((left, right) => left.localeCompare(right))
-}
-
-function buildDateKey(year: number, month: number, day: number) {
-  return new Date(Date.UTC(year, month, day)).toISOString().slice(0, 10)
 }
 
 const PlanningCreateForm = () => {
@@ -27,10 +23,6 @@ const PlanningCreateForm = () => {
   const [eventTitle, setEventTitle] = useState("")
   const [selectedDates, setSelectedDates] = useState<string[]>([])
   const [isPending, startTransition] = useTransition()
-  const [monthCursor, setMonthCursor] = useState(() => {
-    const now = new Date()
-    return { year: now.getUTCFullYear(), month: now.getUTCMonth() }
-  })
 
   const dateFormatter = useMemo(
     () =>
@@ -42,50 +34,6 @@ const PlanningCreateForm = () => {
       }),
     [locale]
   )
-
-  const weekDayFormatter = useMemo(
-    () => new Intl.DateTimeFormat(locale, { weekday: "short" }),
-    [locale]
-  )
-
-  const monthFormatter = useMemo(
-    () => new Intl.DateTimeFormat(locale, { month: "long", year: "numeric" }),
-    [locale]
-  )
-
-  const weekDayLabels = useMemo(() => {
-    return Array.from({ length: 7 }, (_, index) => {
-      const referenceDate = new Date(Date.UTC(2024, 0, 1 + index))
-      return weekDayFormatter.format(referenceDate)
-    })
-  }, [weekDayFormatter])
-
-  const monthLabel = monthFormatter.format(
-    new Date(Date.UTC(monthCursor.year, monthCursor.month, 1))
-  )
-
-  const daysInMonth = new Date(Date.UTC(monthCursor.year, monthCursor.month + 1, 0)).getUTCDate()
-  const firstDayOfMonth = new Date(Date.UTC(monthCursor.year, monthCursor.month, 1)).getUTCDay()
-  const leadingEmptyCells = (firstDayOfMonth + 6) % 7
-  const totalCells = Math.ceil((leadingEmptyCells + daysInMonth) / 7) * 7
-
-  const calendarCells = Array.from({ length: totalCells }, (_, cellIndex) => {
-    const day = cellIndex - leadingEmptyCells + 1
-    if (day < 1 || day > daysInMonth) {
-      return null
-    }
-
-    const dateKey = buildDateKey(monthCursor.year, monthCursor.month, day)
-    return { day, dateKey }
-  })
-
-  const navigateMonth = (direction: "previous" | "next") => {
-    setMonthCursor((previous) => {
-      const monthOffset = direction === "next" ? 1 : -1
-      const nextDate = new Date(Date.UTC(previous.year, previous.month + monthOffset, 1))
-      return { year: nextDate.getUTCFullYear(), month: nextDate.getUTCMonth() }
-    })
-  }
 
   const toggleDate = (dateKey: string) => {
     setSelectedDates((previousDates) => {
@@ -152,57 +100,11 @@ const PlanningCreateForm = () => {
         />
       </div>
 
-      <div className="flex flex-col gap-3">
-        <div className="flex items-center justify-between">
-          <h2 className="text-xl font-semibold">{t("datePickerLabel")}</h2>
-          <div className="flex items-center gap-2">
-            <Button type="button" variant="outline" onClick={() => navigateMonth("previous")} disabled={isPending}>
-              {t("previousMonth")}
-            </Button>
-            <Button type="button" variant="outline" onClick={() => navigateMonth("next")} disabled={isPending}>
-              {t("nextMonth")}
-            </Button>
-          </div>
-        </div>
-
-        <div className="glassPanel flex flex-col gap-3">
-          <p className="text-lg font-semibold capitalize">{monthLabel}</p>
-
-          <div className="grid grid-cols-7 gap-2">
-            {weekDayLabels.map((weekDayLabel) => (
-              <div
-                key={weekDayLabel}
-                className="flex h-8 items-center justify-center text-xs font-semibold text-muted-foreground"
-              >
-                {weekDayLabel}
-              </div>
-            ))}
-          </div>
-
-          <div className="grid grid-cols-7 gap-2">
-            {calendarCells.map((cell, index) => {
-              if (!cell) {
-                return <div key={`empty-${index}`} className="h-10 w-full" />
-              }
-
-              const isSelected = selectedDates.includes(cell.dateKey)
-
-              return (
-                <Button
-                  key={cell.dateKey}
-                  type="button"
-                  variant={isSelected ? "default" : "outline"}
-                  onClick={() => toggleDate(cell.dateKey)}
-                  disabled={isPending}
-                  className={cn("h-10 w-full", isSelected && "font-semibold")}
-                >
-                  {cell.day}
-                </Button>
-              )
-            })}
-          </div>
-        </div>
-      </div>
+      <PlanningDateCalendar
+        selectedDates={selectedDates}
+        onToggleDate={toggleDate}
+        disabled={isPending}
+      />
 
       <div className="flex flex-col gap-3">
         <h2 className="text-xl font-semibold">{t("selectedDatesTitle")}</h2>
